使用单独的 Activity 显示动态详情

+++ bili-api/src/main/kotlin/dev/aaa1115910/biliapi/http/entity/dynamic/DynamicResponse.kt
@@ -40,7 +40,7 @@ data class DynamicItem(
         @SerialName("comment_id_str")
         val commentIdStr: String,
         @SerialName("comment_type")
-        val commentType: Int,
+        val commentType: Long,
         @SerialName("like_icon")
         val likeIcon: LikeIcon,
         @SerialName("rid_str")
