+++ bili-api/src/main/kotlin/dev/aaa1115910/biliapi/http/entity/reply/Comment.kt
@@ -103,7 +103,7 @@ data class CommentData(
         @SerialName("rpid_str")
         val rpidStr: String,
         val state: Int,
-        val type: Int,
+        val type: Long,
         @SerialName("up_action")
         val upAction: UpAction
     ) {
